requirements.txt
arrow==1.2.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cookiecutter==2.4.0
dj-database-url==2.1.0
Django==4.2.5
django-browser-reload==1.11.0
django-tailwind==3.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.2
protobuf==4.21.12
psycopg2-binary==2.9.9
Pygments==2.16.1
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.5.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.8.0
urllib3==2.0.5
whitenoise==6.5.0
